# openapi.yaml
openapi: 3.0.4
info:
  title: Evently
  version: 1.0.12
...
paths:
...
  /api/users/me/created-events:
    get:
      tags:
        - users
      summary: Получить мероприятия, созданные пользователем
      description: Возвращает все мероприятия, созданные пользователя
      operationId: getCreatedEventsForUser
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Invalid ID
        '404':
          description: Event not found
  /api/users/me/attended-events:
    get:
      tags:
        - users
      summary: Получить мероприятия, в которых участвует пользователь
      description: Возвращает все мероприятия, в которых участвует пользователь
      operationId: getAttendedEventsForUser
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Invalid ID
        '404':
          description: Event not found
...
  /api/moderation/users/{userID}:
    get:
      tags:
        - moderation
      summary: Получить пользователя
      description: Возвращает пользователя
      parameters:
        - name: userID
          in: path
          required: true
          example: 42abcd2b-8b9c-4af9-88f7-0bc180cf74b4
          schema:
            type: string
            format: uuid
      operationId: getUserByID
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid ID
        '404':
          description: User not found

    put:
      tags:
        - moderation
      summary: Обновление пользователя модератором
      description: Модератор обновляет данные пользователя
      parameters:
        - name: userID
          in: path
          required: true
          example: 42abcd2b-8b9c-4af9-88f7-0bc180cf74b4
          schema:
            type: string
            format: uuid
      operationId: moderationUpdateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Пользователь успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
        '404':
          description: User not found
        '422':
          description: Validation exception

    delete:
      tags:
        - moderation
      summary: Удалить пользователя по UUID
      description: Удаляет пользователя
      parameters:
        - name: userID
          in: path
          required: true
          example: 42abcd2b-8b9c-4af9-88f7-0bc180cf74b4
          schema:
            type: string
            format: uuid
      operationId: deleteUser
      responses:
        '204':
          description: Successful operation
          content: {}
        '400':
          description: Invalid ID
        '404':
          description: User not found
components:
  schemas:
    Event:
      type: object
      properties:
        event_id:
          type: string
          format: uuid
          example: 42abcd2b-8b9c-4af9-88f7-0bc180cf74b4
        event_title:
          type: string
          example: Карнавал
        description:
          type: string
          example: Очень крутое мероприятие
        start_date:
          description: Дата и время начала
          type: string
          example: 2025-04-10T16:45:00.000Z
        end_date:
          description: Дата и время окончания
          type: string
          example: 2025-04-10T16:45:00.000Z
          nullable: true

        creator:
          $ref: '#/components/schemas/User'
        location:
          type: string
        category:
          $ref: '#/components/schemas/Category'
        participant_count:
          description: Количество участников
          type: integer
          example: 10
          nullable: true
        image_urls:
          type: array
          items:
            type: string
          example:
            - https://example.com/image1.jpg
            - https://example.com/image2.jpg
        event_status:
          type: string
        created_at:
          description: Дата и время окончания
          type: string
          example: 2025-04-10T16:45:00.000Z
          nullable: true
    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          example: 42abcd2b-8b9c-4af9-88f7-0bc180cf74b4
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        nickname:
          type: string
        phone:
          type: string
        profile_pic_url:
          type: string
        user_role:
          type: string
          enum:
            - 'ADMIN'
            - 'USER'
...
    Category:
      type: object
      properties:
        category_id:
          type: string
          format: uuid
        category_name:
          type: string
          example: "Спорт"
...